requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==20.3.0
bcrypt==3.2.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
cryptography==3.4.7
DateTime==4.3
dnspython==2.1.0
ecdsa==0.14.1
environs==9.3.2
fastapi==0.63.0
future==0.18.2
h11==0.12.0
idna==2.10
marshmallow==3.11.1
motor==2.4.0
multidict==5.1.0
passlib==1.7.4
pyArango==1.3.4
pyasn1==0.4.8
pycparser==2.20
pydantic==1.8.1
pymongo==3.11.3
python-dotenv==0.17.0
python-jose==3.2.0
python-multipart==0.0.5
pytz==2021.1
requests==2.25.1
rsa==4.7.2
six==1.15.0
starlette==0.13.6
typing-extensions==3.7.4.3
urllib3==1.26.4
uvicorn==0.13.4
yarl==1.6.3
zope.interface==5.4.0
